% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeptSimilarity.R
\name{adeptSimilarity}
\alias{adeptSimilarity}
\title{Title}
\usage{
adeptSimilarity(x, template.scaled, similarity.measure)
}
\arguments{
\item{similarity.measure}{}
}
\value{

}
\description{
Title
}
\examples{

}
