% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxAndTune.R
\name{finetune_maxima}
\alias{finetune_maxima}
\title{Fine-tune preliminary locations of parttern occurence start and end}
\usage{
finetune_maxima(s.TMP, tau.TMP, nbh.wing, x.Fitted, finetune.maxima.x,
  template.vl.min, template.vl.max)
}
\arguments{
\item{s.TMP}{Index of preliminarily localized parttern occurence start within a signal.}

\item{tau.TMP}{Vector length of preliminarily localized parttern occurence within a signal.}

\item{nbh.wing}{Vector length of each side of a neighbourhood.}

\item{x.Fitted}{Vector of same length as a signal length; contains information
whether or not a signal has a pattern occurence already identified at certain index.}

\item{finetune.maxima.x}{Singal (or its smoothed version) within which we are
searching for pattern occurrences.}

\item{template.vl.min}{Minimal assumed vector length of a pattern occurrence we are searching for.}

\item{template.vl.max}{Maximal assumed vector length of a pattern occurrence we are searching for.}
}
\value{
2-element vector of tuned: (1) index of localized parttern occurence start within a signal,
(2) vector length of localized parttern occurence within a signal.
}
\description{
Fine-tune preliminary locations of parttern occurence start and end
via identifying singal peaks in respective neighbourhoods of
occurence start and end locations.
}
