% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningWinSmooth.R
\name{get.x.smoothed}
\alias{get.x.smoothed}
\title{Wrapper function for \code{RunningWinSmooth}}
\usage{
get.x.smoothed(x, W, x.fs = 1, NA.repl.surce.k = 4)
}
\arguments{
\item{x}{Numeric vector to be smoothed.}

\item{W}{A width of a moving window given in time (seconds).}

\item{x.fs}{Frequency of \code{x} expressed in number of observations collected
per second. Defaults to \code{1}.}

\item{NA.repl.surce.k}{Scalar for number of subsequent/procceeding non-\code{NA} values that appear
in smoothed signal and are used to provide sample mean replacement for \code{NA} values.}
}
\value{
Smoothed signal with  \code{NA} values replaced with sample means of
subsequent/procceeding non-\code{NA} values that appear
in smoothed signal.
}
\description{
Wrapper function for \code{RunningWinSmooth}. Replaces \code{NA} values with
which appear in head and tail of smoothed signal as a result of MA not defined
for first and last \code{k} elements of a signal. The \code{NA} values are replaced
with sample means of subsequent/procceeding non-\code{NA} values that appear
in smoothed signal.
}
\examples{
set.seed(20191021)
x <- sin(seq(0, 2 * pi, length.out = 1000)) + rnorm(1000, sd = 0.2)
x.smoothed <- get.x.smoothed(x, 20)
plot(x, type = "l", col = "grey")
lines(x.smoothed, col = "blue")

}
