% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My hello world function}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of a person to say hi to}
}
\value{
The output from \code{\link{print}}
}
\description{
My hello world function
}
\examples{
hello("Marta")
\dontrun{
hello("Steve")
}
}
