% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxAndTune.R
\name{maxAndTune}
\alias{maxAndTune}
\title{Title}
\usage{
maxAndTune(x, template.vl, similarity.mat, similarity.measure.thresh,
  finetune, finetune.maxima.x, finetune.maxima.nbh.vl)
}
\arguments{
\item{finetune.maxima.nbh.vl}{}
}
\description{
Title
}
