% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{Silly Printer Function}
\usage{
printer(r, x)
}
\arguments{
\item{r}{What you want in the second column}

\item{x}{What you want in the first column}
}
\value{
tibble
}
\description{
Silly Printer Function
}
\examples{
printer(x = rnorm(5), r = rnorm(5))
}
