% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeptSimilarity.R
\name{adeptSimilarity}
\alias{adeptSimilarity}
\title{Title}
\usage{
adeptSimilarity(x, template.scaled, similarity.measure,
  run.parallel = FALSE, run.parallel.ncores = NULL)
}
\arguments{
\item{run.parallel.ncores}{}
}
\value{

}
\description{
Title
}
\examples{

}
