% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeptSimilarity.R
\name{adeptSimilarity}
\alias{adeptSimilarity}
\title{ADEPT Similarity Matrix Computation}
\usage{
adeptSimilarity(x, template.scaled, similarity.measure)
}
\arguments{
\item{x}{A numerical vector. A time-series \code{x}.}

\item{template.scaled}{A list of lists of numeric vectors.  Each element of
\code{template.scaled} list is a list of pattern templates scaled according
to a particular scale parameter (that is, scaled to a particular vector length).
Number of elements in the \code{template.scaled} list corresponds to the number
 of unique scale values considered in the method. See: \code{scaleTemplate {adept}}
 for computation of \code{template.scaled} list.}

\item{similarity.measure}{A character scalar. Defines a statistic
used in similarity matrix computation. One of the following:
\itemize{
  \item "cov" - for covariance,
  \item "cor" - for correlation.
}}
}
\value{
A numeric matrix. The matrix entries are similarity values between
time-series \code{x} windows and a collection
of scaled versions of pattern templates. Each matrix row consists
of a vector of similarity statistic between \code{x} windows and a pattern
rescaled to a particular scale parameter (that is, scaled to a particular vector length);
precisely, for a scale-specific matrix row vector, each vector element is a maximum
similarity value selected out of all distinct pattern templates considered.
}
\description{
Compute ADEPT similarity matrix between time-series \code{x} windows and a collection
of scaled versions of pattern templates.
}
\examples{
## Simulate data
par(mfrow = c(1,1))
x0 <- sin(seq(0, 2 * pi * 100, length.out = 10000))
x  <- x0 + rnorm(1000, sd = 0.1)
template <- list(x0[1:500])
template.vl <- seq(300, 700, by = 50)

## Rescale pattern
template.scaled <- scaleTemplate(template, template.vl)

## Compute ADEPT similarity matrix
out <- adeptSimilarity(x, template.scaled, "cov")

\dontrun{
## Visualize
par(mfrow = c(1,1))
image(t(out),
      main = "ADEPT similarity matrix\\nfor time-series x and scaled versions of a pattern",
      xlab = "Time-series x index",
      ylab = "Pattern scale (pattern vector length)",
      xaxt = "n", yaxt = "n")
xaxis <- c(1, seq(1000, length(x0), by = 1000))
yaxis <- template.vl
axis(1, at = xaxis/max(xaxis), labels = xaxis)
axis(2, at = (yaxis - min(yaxis))/(max(yaxis) - min(yaxis)), labels = yaxis)
}

}
\seealso{
\code{scaleTemplate {adept}}
}
