% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleTemplate.R
\name{scaleTemplate}
\alias{scaleTemplate}
\title{Compute a list of rescaled versions of empirical pattern(s)}
\usage{
scaleTemplate(template, template.vl)
}
\arguments{
\item{template}{A list of numerical vector(s) representing
empirical pattern(s).}

\item{template.vl}{A numeric vector of vector lengths to which each of
the empirical pattern(s) will be scaled to.}
}
\value{
A list of scaled versions of empirical pattern(s).
Number of elements in the list corresponds to the number of unique pattern scale
values considered. Each element of the list is a list
itself that contains a numeric vector(s), where each vector represents
one of possibly multiple distinct empirical patterns considered, scaled
to have the same scale value (same vector length).
}
\description{
Compute a list of rescaled versions of empirical pattern(s) via linear interpolation.
}
\examples{
template <- list(sin(seq(0, 2 * pi, length.out = 100)),
                 cos(seq(0, 2 * pi, length.out = 100)))
template.vl <- c(50, 100, 200)
out <- scaleTemplate(template, template.vl)
par(mfrow = c(2, 1))
plot(out[[3]][[1]], type = "l",
     main = "pattern: sin([0,2*pi]); different scales considered",
     ylab = "pattern", xlab = "output vector index")
lines(out[[2]][[1]], col = "red")
lines(out[[1]][[1]], col = "blue")
plot(out[[3]][[2]], type = "l",
     main = "pattern: cos([0,2*pi]); different scales considered",
     ylab = "pattern", xlab = "output vector index")
lines(out[[2]][[2]], col = "red")
lines(out[[1]][[2]], col = "blue")

}
