% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentPattern.R
\name{segmentPattern}
\alias{segmentPattern}
\title{Title}
\usage{
segmentPattern(x, x.fs, template, pattern.dur.seq = NULL,
  similarity.measure = "cov", similarity.measure.thresh = 0,
  x.adept.ma.W = NULL, finetune = NULL, finetune.maxima.ma.W = NULL,
  finetune.maxima.nbh.W = NULL, run.parallel = FALSE,
  run.parallel.ncores = NULL)
}
\arguments{
\item{run.parallel.ncores}{}
}
\value{
Data frame with summary of segmented pattern occurences; has the
following columns:
\describe{
  \item{One}{First item}
  \item{Two}{Second item}
}
}
\description{
Title
}
