% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentPattern.R
\name{segmentPattern}
\alias{segmentPattern}
\title{Pattern Segmentation From a Time-series via ADEPT}
\usage{
segmentPattern(x, x.fs, template, pattern.dur.seq,
  similarity.measure = "cov", similarity.measure.thresh = 0,
  x.adept.ma.W = NULL, finetune = NULL, finetune.maxima.ma.W = NULL,
  finetune.maxima.nbh.W = NULL, run.parallel = FALSE,
  run.parallel.cores = NULL, x.cut = TRUE, x.cut.vl = 6000)
}
\arguments{
\item{x}{A numeric vector. A time-series to segment pattern occurrences from.}

\item{x.fs}{A numeric scalar. Frequency at which a time-series \code{x} is collected,
expressed in a number of observations per second.}

\item{template}{A list of numeric vectors, or a numeric vector.
Each vector represents a distinct pattern template used in segmentation.}

\item{pattern.dur.seq}{A numeric vector. A grid of pattern duration
times used in segmentation. Expressed in seconds. See: Details.}

\item{similarity.measure}{A character scalar. Statistic used to compute similarity
between a time-series \code{x}  and pattern templates. Currently supported values:
\itemize{
  \item \code{"cov"} - covariance,
  \item \code{"cor"} - correlation,
}
Default is \code{"cov"}.}

\item{similarity.measure.thresh}{A numeric scalar. Threshold of minimal similarity
value between a time-series \code{x} and pattern templates
below which the algorithm does not identify a pattern occurrence.
Default is \code{0}.}

\item{x.adept.ma.W}{A numeric scalar.
A length of a window used in moving average smoothing of a time-series \code{x} for
  similarity matrix computation. Expressed in seconds.
 Default is \code{NULL} (no smoothing applied).}

\item{finetune}{A character scalar. A type of fine-tuning procedure empolyed in
 segmentation. Defaults to \code{NULL} (no ine-tuning procedure empolyed). Currently supported values:
\itemize{
  \item \code{"maxima"} - tunes preliminarily identified locations of parttern occurence
  beginning and end so
  as they correspond to local maxima of time-series \code{x} (or smoothed version of \code{x})
  found within neighbourhoods of preliminary locations.
}}

\item{finetune.maxima.ma.W}{A numeric scalar.
A length of a window used in moving average smoothing of a time-series \code{x} in
 \code{"maxima"} fine-tuning procedure. Expressed in seconds.
 Default is \code{NULL} (no smoothing applied).}

\item{finetune.maxima.nbh.W}{A numeric scalar.
A length of the two neighborhoods centered at preliminarily identified pattern occurrence beginning and end points
within which we search for local maxima of \code{x} (or smoothed version of \code{x}) in \code{"maxima"}
fine-tuning procedure. Expressed in seconds. Default is \code{NULL}.
Note: if the length provided corresponds to an even number of \code{x} vector indices,
it will be rounded down so as the corresponding number of vector indices is its closest odd number.}

\item{run.parallel}{A logical scalar. Whether or not to use parallel execution in the algorithm.
The \code{future} package
is used to execute code asynchronously. Default is \code{FALSE}.}

\item{run.parallel.cores}{An integer scalar.
The number of cores to use for parallel execution.
Default is \code{NULL}. If not specified, the number of cores is set to a number of
cores available minus 1.}

\item{x.cut}{A logical scalar. Whether or not to use time optimization procedure in
which a time-series \code{x} is cut into parts and segmentation is performed for
each part of \code{x} separately. Recommended for a time-series \code{x} of vector length
 above 30,000. Default is \code{TRUE}.}

\item{x.cut.vl}{An integer scalar.
Defines a vector length of parts that \code{x} vector is cut into during the execution time optimization procedure.
Default is \code{6000} (recommended).}
}
\value{
A \code{data.frame} with segmentation results. Each row
of the returned \code{data.frame} describes one identified pattern occurrence:
\itemize{
  \item \code{tau_i} - index of a time-series \code{x} where identified pattern occurence starts,
  \item \code{T_i} - duration of identified pattern occurence starts, expressed in  a time-series \code{x} vector length,
  \item \code{sim_i} -  value of similarity statistic between an identified pattern occurence and corresponding
  window of a time-series used in similarity matrix computation;
  note: this value corresponds to similarity statistic between
  preliminarily identified pattern occurence and corresponding window of a time-series used in similarity matrix computation;
  specifically: if the fine-tune procedure is employed,
  the similarity value between the final pattern occurence location and corresponding window of time-series \code{x}
  singal may differ from the value in this table.
}
}
\description{
Segment pattern from a time-series \code{x} via Adaptive Empirical Pattern
Transformation (ADEPT).
}
\details{
Function implements Adaptive Empirical Pattern Transformation (ADEPT) method for pattern segmentation
from a time-series \code{x}.
ADEPT was designed with the aim of performing fast, accurate walking strides segmentation
from high-density data
collected from wearable accelerometer worn during continuous walking activity.

    ADEPT identifies pattern occurrenes from a time-series \code{x} via maximizing similarity
(correlation, covariance etc.) between a time-series \code{x} and pattern
templates.
\itemize{
  \item Pattern template is scaled to various scale parameters to allow
  for a potentially better match with pattern occurrences that may change
  its duration over time.  \code{pattern.dur.seq} argument defines a grid of a
   pattern duration to be used in segmentation. Note:  the more dense grid may potentially
   increase segmentation accuracy but may also increase method execution time.
  \item Multiple pattern templates are allowed simultaneously to
  allow for potentially better match with pattern occurrences
  that may change its shape over time.
}
In practice, a pre-defined pattern template may be derived as an empirical pattern, that is,
a data-derived vector representing a pattern of interest.
}
\examples{




}
